requirements.txt
streamlit>=1.35.0
groq>=0.9.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
